Rename ADVI.py to advi.py.
