+++ pymc3/variational/__init__.py
@@ -1,1 +1,1 @@
-from .ADVI import advi
+from .advi import advi
